requirements.txt
flask
tensorflow==2.20.0
numpy
pillow
gunicorn